Vuetify migration: replace Flot with uPlot for the timeline graph
Phase 2b step 1: drop Flot rendering entirely. Visualizations.vue
delegates the timeline to a new TimelineGraph.vue component built on
uPlot. The POC (TimelineUplotPoc.vue) is removed since TimelineGraph
is now the production code.

What's gone:
- Three jQuery Flot scripts from viewer/public/ (flot.min, event.drag,
  flot.resize) ~55 KB total
- The dynamic 'public/jquery.flot.*' imports + the $.plot init assertion
  from Visualizations.vue's created() hook
- Visualizations.vue methods setupGraphElement, setupGraphData,
  addBusinessHours, calculateHoverBarWidth (~260 lines)
- Watchers on graphType / seriesType / capStartTimes that re-ran the Flot
  setup; TimelineGraph reacts to its own props instead
- Module-level color/bar-width lets that only fed the Flot options
- The plotArea / plot / plotCheck / graph / graphOptions / initialized
  data fields
- The orphaned Sessions.vue:649 superstition comment was already gone

What's new:
- TimelineGraph.vue (443 lines): uPlot-based renderer with theme colors
  read from CSS custom props, capture-restart vertical markers, business-
  hours overlay bands, drag-to-select for time range, custom tooltip with
  the Flot 'X out of Y total filtered <type> on <date>' format, and the
  adaptive client-side rebucketing that came out of the POC tuning.
- Public methods panLeft / panRight / zoomIn / zoomOut / getXRange called
  from Visualizations.vue's existing pan/zoom buttons via $ref.

jvectormap is still loaded — the world map swap to D3+topojson lands in
the next 2b commit. WorldMapD3Poc + ?vizpoc=d3map remain available for
side-by-side comparison.
